+++ samples/My.Hr/My.Hr.Test/Validators/PerformanceReviewValidatorTest.cs
@@ -1,10 +1,11 @@
 ﻿using Beef.Test.NUnit;
 using Beef.Validation;
+using Microsoft.Extensions.DependencyInjection;
 using Moq;
 using My.Hr.Business;
-using My.Hr.Business.Validation;
-using My.Hr.Common.Agents;
-using My.Hr.Common.Entities;
+using My.Hr.Business.Data;
+using My.Hr.Business.DataSvc;
+using My.Hr.Business.Entities;
 using NUnit.Framework;
 using System;
 using System.Threading.Tasks;
@@ -14,30 +15,36 @@ namespace My.Hr.Test.Validators
     [TestFixture]
     public class PerformanceReviewValidatorTest
     {
-        private readonly Mock<IReferenceDataAgent> _referenceData = new Mock<IReferenceDataAgent>();
-        private readonly Mock<IEmployeeManager> _employeeManager = new Mock<IEmployeeManager>();
-        private readonly Mock<IPerformanceReviewManager> _perfReviewManager = new Mock<IPerformanceReviewManager>();
+        private Func<IServiceCollection, IServiceCollection>? _testSetup;
 
         [OneTimeSetUp]
         public void OneTimeSetUp()
         {
-            _referenceData.Setup(x => x.PerformanceOutcomeGetAllAsync(null, null)).ReturnsWebApiAgentResultAsync(new PerformanceOutcomeCollection { new PerformanceOutcome { Id = Guid.NewGuid(), Code = "ME" } });
+            var rd = new Mock<IReferenceDataData>();
+            rd.Setup(x => x.PerformanceOutcomeGetAllAsync()).ReturnsAsync(new PerformanceOutcomeCollection { new PerformanceOutcome { Id = Guid.NewGuid(), Code = "ME" } });
 
-            _employeeManager.Setup(x => x.GetAsync(404.ToGuid())).ReturnsAsync((Employee)null!);
-            _employeeManager.Setup(x => x.GetAsync(1.ToGuid())).ReturnsAsync(new Employee { Id = 1.ToGuid(), StartDate = DateTime.Now.AddYears(-1) });
-            _employeeManager.Setup(x => x.GetAsync(2.ToGuid())).ReturnsAsync(new Employee { Id = 2.ToGuid(), StartDate = DateTime.Now.AddYears(-1), Termination = new TerminationDetail { Date = DateTime.Now.AddMonths(-1) } });
+            var em = new Mock<IEmployeeManager>();
+            em.Setup(x => x.GetAsync(404.ToGuid())).ReturnsAsync((Employee)null!);
+            em.Setup(x => x.GetAsync(1.ToGuid())).ReturnsAsync(new Employee { Id = 1.ToGuid(), StartDate = DateTime.Now.AddYears(-1) });
+            em.Setup(x => x.GetAsync(2.ToGuid())).ReturnsAsync(new Employee { Id = 2.ToGuid(), StartDate = DateTime.Now.AddYears(-1), Termination = new TerminationDetail { Date = DateTime.Now.AddMonths(-1) } });
 
-            _perfReviewManager.Setup(x => x.GetAsync(1.ToGuid())).ReturnsAsync(new PerformanceReview { Id = 1.ToGuid(), EmployeeId = 2.ToGuid() });
+            var prm = new Mock<IPerformanceReviewManager>();
+            prm.Setup(x => x.GetAsync(1.ToGuid())).ReturnsAsync(new PerformanceReview { Id = 1.ToGuid(), EmployeeId = 2.ToGuid() });
+
+            _testSetup = sc => sc
+                .AddGeneratedValidationServices()
+                .AddGeneratedReferenceDataManagerServices()
+                .AddGeneratedReferenceDataDataSvcServices()
+                .AddScoped(_ => rd.Object)
+                .AddScoped(_ => em.Object)
+                .AddScoped(_ => prm.Object);
         }
 
         [Test]
         public async Task A110_Validate_Initial()
         {
             await ValidationTester.Test()
-                .ConfigureServices(ServiceCollectionsValidationExtension.AddGeneratedValidationServices)
-                .AddScopedService(_referenceData)
-                .AddScopedService(_employeeManager)
-                .AddScopedService(_perfReviewManager)
+                .ConfigureServices(_testSetup!)
                 .ExpectMessages(
                     "Employee is required.",
                     "Date is required.",
@@ -59,10 +66,7 @@ namespace My.Hr.Test.Validators
             };
 
             await ValidationTester.Test()
-                .ConfigureServices(ServiceCollectionsValidationExtension.AddGeneratedValidationServices)
-                .AddScopedService(_referenceData)
-                .AddScopedService(_employeeManager)
-                .AddScopedService(_perfReviewManager)
+                .ConfigureServices(_testSetup!)
                 .ExpectMessages(
                     "Date must be less than or equal to today.",
                     "Outcome is invalid.",
@@ -85,10 +89,7 @@ namespace My.Hr.Test.Validators
             };
 
             await ValidationTester.Test()
-                .ConfigureServices(ServiceCollectionsValidationExtension.AddGeneratedValidationServices)
-                .AddScopedService(_referenceData)
-                .AddScopedService(_employeeManager)
-                .AddScopedService(_perfReviewManager)
+                .ConfigureServices(_testSetup!)
                 .ExpectMessages("Date must not be prior to the Employee starting.")
                 .CreateAndRunAsync<IValidator<PerformanceReview>, PerformanceReview>(pr);
         }
@@ -106,10 +107,7 @@ namespace My.Hr.Test.Validators
             };
 
             await ValidationTester.Test()
-                .ConfigureServices(ServiceCollectionsValidationExtension.AddGeneratedValidationServices)
-                .AddScopedService(_referenceData)
-                .AddScopedService(_employeeManager)
-                .AddScopedService(_perfReviewManager)
+                .ConfigureServices(_testSetup!)
                 .ExpectMessages("Date must not be after the Employee has terminated.")
                 .CreateAndRunAsync<IValidator<PerformanceReview>, PerformanceReview>(pr);
         }
@@ -129,11 +127,8 @@ namespace My.Hr.Test.Validators
 
             // Need to set the OperationType to Update to exercise logic.
             await ValidationTester.Test()
+                .ConfigureServices(_testSetup!)
                 .OperationType(Beef.OperationType.Update)
-                .ConfigureServices(ServiceCollectionsValidationExtension.AddGeneratedValidationServices)
-                .AddScopedService(_referenceData)
-                .AddScopedService(_employeeManager)
-                .AddScopedService(_perfReviewManager)
                 .ExpectErrorType(Beef.ErrorType.NotFoundError)
                 .CreateAndRunAsync<IValidator<PerformanceReview>, PerformanceReview>(pr);
         }
@@ -153,11 +148,8 @@ namespace My.Hr.Test.Validators
 
             // Need to set the OperationType to Update to exercise logic.
             await ValidationTester.Test()
+                .ConfigureServices(_testSetup!)
                 .OperationType(Beef.OperationType.Update)
-                .ConfigureServices(ServiceCollectionsValidationExtension.AddGeneratedValidationServices)
-                .AddScopedService(_referenceData)
-                .AddScopedService(_employeeManager)
-                .AddScopedService(_perfReviewManager)
                 .ExpectMessages("Employee is not allowed to change; please reset value.")
                 .CreateAndRunAsync<IValidator<PerformanceReview>, PerformanceReview>(pr);
         }
@@ -176,11 +168,8 @@ namespace My.Hr.Test.Validators
 
             // Need to set the OperationType to Create to exercise logic.
             await ValidationTester.Test()
+                .ConfigureServices(_testSetup!)
                 .OperationType(Beef.OperationType.Create)
-                .ConfigureServices(ServiceCollectionsValidationExtension.AddGeneratedValidationServices)
-                .AddScopedService(_referenceData)
-                .AddScopedService(_employeeManager)
-                .AddScopedService(_perfReviewManager)
                 .CreateAndRunAsync<IValidator<PerformanceReview>, PerformanceReview>(pr);
         }
 
@@ -199,11 +188,8 @@ namespace My.Hr.Test.Validators
 
             // Need to set the OperationType to Update to exercise logic.
             await ValidationTester.Test()
+                .ConfigureServices(_testSetup!)
                 .OperationType(Beef.OperationType.Update)
-                .ConfigureServices(ServiceCollectionsValidationExtension.AddGeneratedValidationServices)
-                .AddScopedService(_referenceData)
-                .AddScopedService(_employeeManager)
-                .AddScopedService(_perfReviewManager)
                 .CreateAndRunAsync<IValidator<PerformanceReview>, PerformanceReview>(pr);
         }
     }